requirements.txt
fastapi>=0.104.0
pydantic>=2.5.0
uvicorn>=0.24.0
neo4j>=5.15.0
python-dotenv>=1.0.0
numpy>=1.24.3
faiss-cpu>=1.7.4
tiktoken>=0.5.1
torch>=2.2.0
transformers>=4.40.0
sentence-transformers>=2.2.2
openai>=1.6.0
rapidfuzz>=3.5.0
networkx>=3.2
diskcache>=5.6.1
prometheus-fastapi-instrumentator>=6.1.0
sse-starlette>=1.6.5
jsonschema>=4.20.0
tenacity>=8.2.3
jinja2>=3.1.2
pytest>=7.4.0

# MLX dependencies (Apple Silicon only)
mlx>=0.0.6; platform_machine == "arm64"
mlx-lm>=0.0.6; platform_machine == "arm64"

# Information Extraction Service
transformers>=4.40.0
torch>=2.2.0
sentencepiece>=0.1.97
protobuf>=3.20.0
requests>=2.28.0

# === NER Dependencies ===
spacy>=3.7.3
tner>=0.0.9
# Download spacy model: python -m spacy download en_core_web_sm
rich